requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
lxml==4.9.3
mysql-connector-python==8.2.0
protobuf==4.21.12
requests==2.31.0
roman==4.1
soupsieve==2.5
urllib3==2.0.7
